Move test class out of tofix package for #4629
