+++ src/test/java/tools/jackson/databind/records/RecordsWithJsonIncludeAndIgnorals4629Test.java
@@ -1,4 +1,4 @@
-package tools.jackson.databind.records.tofix;
+package tools.jackson.databind.records;
 
 import org.junit.jupiter.api.Test;
 
@@ -7,7 +7,6 @@ import com.fasterxml.jackson.annotation.JsonIncludeProperties;
 
 import tools.jackson.databind.ObjectMapper;
 import tools.jackson.databind.testutil.DatabindTestUtil;
-import tools.jackson.databind.testutil.failure.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
